reorganizing; cleanup; Fixed some warnings
