+++ Core/Inc/robot.h
@@ -1,0 +1,3 @@
+
+void init(void);
+void loop(void);
